requirements.txt
absl-py==0.9.0
apptools==4.5.0
arch==4.15
arrow==0.15.8
astor==0.8.1
attrs==19.3.0
autopep8==1.5
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.1
bleach==3.1.5
brotlipy==0.7.0
bs4==0.0.1
cachetools==4.0.0
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
conda==4.9.2
conda-build==3.18.8
conda-package-handling==1.7.0
conda-verify==3.4.2
configobj==5.0.6
cryptography==2.9.2
cv==1.0.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
future==0.18.2
gast==0.2.2
get==2019.4.13
glob2==0.7
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5json==1.1.3
imageio==2.9.0
imgviz==1.2.1
ipyevents==0.8.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itchat==1.3.10
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jupyter-core==4.6.3
jupyterlab==2.1.5
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
labelImg==1.8.3
labelme==4.5.6
libarchive-c==2.9
llvmlite==0.34.0
lxml==4.5.2
Markdown==3.2.1
matplotlib==3.3.0
mayavi==4.7.1
menuinst==1.4.16
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nbformat==5.0.7
networkx==2.5
notebook==6.0.3
numba==0.51.2
oauthlib==3.1.0
open3d==0.8.0.0
open3d-python==0.7.0.0
opencv-python==4.2.0.32
opt-einsum==3.1.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
Pillow==7.2.0
pipreqs==0.4.10
pkginfo==1.5.0.1
plyfile==0.7.1
post==2019.4.13
prometheus-client==0.8.0
prompt-toolkit==3.0.5
property-cached==1.6.4
protobuf==3.11.3
psutil==5.7.0
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
Pygments==2.6.1
PyOpenGL==3.1.5
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyrsistent==0.16.0
PySnooper==0.3.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QtPy==1.9.0
query-string==2019.4.13
request==2019.4.13
requests-oauthlib==1.3.0
rsa==4.0
ruamel-yaml==0.15.87
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.2.1
seaborn==0.10.1
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
statsmodels==0.12.0
tensorboard==2.0.2
tensorboardX==2.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.2.0
TorchSnooper==0.8
torchsummary==1.5.1
torchvision==0.4.1
tornado==6.0.4
traitlets==4.3.3
tushare==1.2.60
typed-ast==1.4.1
urllib3==1.25.9
vtk===-PKG-VERSION
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
xlrd==1.2.0
yapf==0.29.0
yarg==0.1.9
yarn==0.0.9
zipp==3.1.0
